client/server: Reimplement using xmlrpc
Twisted's reactor is meant for clients that have an event-loop.  The
client that devlib wants to use has to return, but the reactor is not
meant to be restartable.  execute_command() tried to workaround it by
uninstalling it and reinstalling it, but that lead to resource leaks
that weren't easy to patch.  With python 3.8 and twisted 19.10.0 you
were upsetting the signal handler:

  Exception ignored when trying to write to the signal wakeup fd:
  Traceback (most recent call last):
    File "twisted/internet/_signals.py", line 56, in noopSignalHandler
      def noopSignalHandler(*args):
  OSError: [Errno 9] Bad file descriptor

This sometimes lead to python crashing.  Reimplement using xmlrpc to
get rid of twisted, and simplify the code accordingly.

Fixes #10